requirements.txt
# FastAPI and ASGI server
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# Database and ORM
sqlalchemy>=2.0.0
sqlmodel>=0.0.8
alembic>=1.10.0

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5

# Validation and serialization
pydantic>=2.0.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0

# Templates and Static Files
jinja2>=3.1.0
aiofiles>=23.0.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
httpx>=0.24.0

# HTTP client for proxy
httpx>=0.24.0

# Database drivers
# SQLite (built-in)
# For PostgreSQL: uncomment the line below
# psycopg2-binary>=2.9.0

# Utilities
python-dotenv>=1.0.0
